backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "motor>=3.3.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.111.0",
    "beanie>=1.30.0",
    "boto3>=1.34.0",
    "types-boto3[essential]>=1.39.8",
    "uvicorn[standard]>=0.35.0",
    "openai>=1.97.0",
    "marker-pdf>=0.2.0",
]

[dependency-groups]
dev = [
    "types-boto3>=1.39.8",
]
